Move blueoil/nn/quantizations to blueoil/quantizations
