+++ blueoil/configs/core/classification/darknet_quantize_cifar10.py
@@ -29,7 +29,7 @@ from blueoil.data_augmentor import (
     FlipLeftRight,
     Pad,
 )
-from blueoil.nn.quantizations import (
+from blueoil.quantizations import (
     binary_channel_wise_mean_scaling_quantizer,
     linear_mid_tread_half_quantizer,
 )
