+++ blueoil/configs/core/classification/darknet_quantize_ilsvrc_2012.py
@@ -32,7 +32,7 @@ from blueoil.pre_processor import (
     Resize,
     DivideBy255,
 )
-from blueoil.nn.quantizations import (
+from blueoil.quantizations import (
     binary_channel_wise_mean_scaling_quantizer,
     linear_mid_tread_half_quantizer,
 )
