+++ blueoil/configs/core/classification/lm_resnet_quantize_cifar10.py
@@ -30,7 +30,7 @@ from blueoil.data_augmentor import (
     Pad,
 )
 
-from blueoil.nn.quantizations import (
+from blueoil.quantizations import (
     binary_mean_scaling_quantizer,
     linear_mid_tread_half_quantizer,
 )
