+++ blueoil/configs/core/keypoint_detection/lm_single_pose_v1_quantize_mscoco.py
@@ -33,7 +33,7 @@ from blueoil.data_augmentor import (
     Color,
     Contrast
 )
-from blueoil.nn.quantizations import (
+from blueoil.quantizations import (
     binary_channel_wise_mean_scaling_quantizer,
     linear_mid_tread_half_quantizer,
 )
