+++ blueoil/configs/core/object_detection/lm_fyolo_quantize_pascalvoc_2007_2012.py
@@ -40,7 +40,7 @@ from blueoil.data_augmentor import (
     SSDRandomCrop,
 )
 
-from blueoil.nn.quantizations import (
+from blueoil.quantizations import (
     binary_channel_wise_mean_scaling_quantizer,
     linear_mid_tread_half_quantizer,
 )
