+++ blueoil/configs/core/segmentation/lm_bisenet_quantize_camvid.py
@@ -35,7 +35,7 @@ from blueoil.data_augmentor import (
     FlipLeftRight,
     Hue,
 )
-from blueoil.nn.quantizations import (
+from blueoil.quantizations import (
     binary_channel_wise_mean_scaling_quantizer,
     linear_mid_tread_half_quantizer,
 )
