+++ blueoil/configs/core/segmentation/lm_segnet_v1_quantize_camvid.py
@@ -31,7 +31,7 @@ from blueoil.data_augmentor import (
     FlipLeftRight,
     Hue,
 )
-from blueoil.nn.quantizations import (
+from blueoil.quantizations import (
     binary_mean_scaling_quantizer,
     linear_mid_tread_half_quantizer,
 )
