+++ blueoil/configs/core/segmentation/segnet_quantize_bdd100k.py
@@ -23,7 +23,7 @@ from blueoil.data_processor import Sequence
 from blueoil.networks.segmentation.lm_segnet_quantize import LmSegnetQuantize
 from blueoil.datasets.bdd100k import BDD100KSegmentation
 from blueoil.pre_processor import DivideBy255, Resize
-from blueoil.nn.quantizations import (binary_mean_scaling_quantizer,
+from blueoil.quantizations import (binary_mean_scaling_quantizer,
                                       linear_mid_tread_half_quantizer)
 
 IS_DEBUG = False
