+++ blueoil/configs/core/segmentation/segnet_quantize_cityscapes.py
@@ -31,7 +31,7 @@ from blueoil.data_augmentor import (
     FlipLeftRight,
     Hue,
 )
-from blueoil.nn.quantizations import (
+from blueoil.quantizations import (
     binary_mean_scaling_quantizer,
     linear_mid_tread_half_quantizer,
 )
