+++ blueoil/configs/example/object_detection.py
@@ -37,7 +37,7 @@ from blueoil.data_augmentor import (
     Hue,
     SSDRandomCrop,
 )
-from blueoil.nn.quantizations import (
+from blueoil.quantizations import (
     binary_channel_wise_mean_scaling_quantizer,
     linear_mid_tread_half_quantizer,
 )
