+++ blueoil/networks/classification/lmnet_v1.py
@@ -118,9 +118,9 @@ class LmnetV1Quantize(LmnetV1):
     ``weight_quantizer``, ``weight_quantizer_kwargs``.
 
     Args:
-        activation_quantizer (callable): Weight quantizater. See more at `blueoil.nn.quantizations`.
+        activation_quantizer (callable): Weight quantizater. See more at `blueoil.quantizations`.
         activation_quantizer_kwargs (dict): Kwargs for `activation_quantizer`.
-        weight_quantizer (callable): Activation quantizater. See more at `blueoil.nn.quantizations`.
+        weight_quantizer (callable): Activation quantizater. See more at `blueoil.quantizations`.
         weight_quantizer_kwargs (dict): Kwargs for `weight_quantizer`.
     """
     version = 1.0
