+++ blueoil/networks/keypoint_detection/lm_single_pose_v1.py
@@ -144,9 +144,9 @@ class LmSinglePoseV1Quantize(LmSinglePoseV1):
     ``weight_quantizer``, ``weight_quantizer_kwargs``.
 
     Args:
-        activation_quantizer (callable): Weight quantizater. See more at `blueoil.nn.quantizations`.
+        activation_quantizer (callable): Weight quantizater. See more at `blueoil.quantizations`.
         activation_quantizer_kwargs (dict): Kwargs for `activation_quantizer`.
-        weight_quantizer (callable): Activation quantizater. See more at `blueoil.nn.quantizations`.
+        weight_quantizer (callable): Activation quantizater. See more at `blueoil.quantizations`.
         weight_quantizer_kwargs (dict): Kwargs for `weight_quantizer`.
 
     """
