+++ blueoil/networks/segmentation/lm_bisenet.py
@@ -363,9 +363,9 @@ class LMBiSeNetQuantize(LMBiSeNet):
        ``weight_quantizer``, ``weight_quantizer_kwargs``.
 
     Args:
-        activation_quantizer (callable): Weight quantizer. See more at `blueoil.nn.quantizations`.
+        activation_quantizer (callable): Weight quantizer. See more at `blueoil.quantizations`.
         activation_quantizer_kwargs (dict): Kwargs for `activation_quantizer`.
-        weight_quantizer (callable): Activation quantizer. See more at `blueoil.nn.quantizations`.
+        weight_quantizer (callable): Activation quantizer. See more at `blueoil.quantizations`.
         weight_quantizer_kwargs (dict): Kwargs for `weight_quantizer`.
     """
