+++ blueoil/networks/segmentation/lm_segnet_v0.py
@@ -109,9 +109,9 @@ class LmSegnetV0Quantize(LmSegnetV0):
     ``weight_quantizer``, ``weight_quantizer_kwargs``.
 
     Args:
-        activation_quantizer (callable): Weight quantizater. See more at `blueoil.nn.quantizations`.
+        activation_quantizer (callable): Weight quantizater. See more at `blueoil.quantizations`.
         activation_quantizer_kwargs (dict): Kwargs for `activation_quantizer`.
-        weight_quantizer (callable): Activation quantizater. See more at `blueoil.nn.quantizations`.
+        weight_quantizer (callable): Activation quantizater. See more at `blueoil.quantizations`.
         weight_quantizer_kwargs (dict): Kwargs for `weight_quantizer`.
     """
