+++ blueoil/networks/segmentation/lm_segnet_v1.py
@@ -99,10 +99,10 @@ class LmSegnetV1Quantize(LmSegnetV1):
 
     Args:
         activation_quantizer (callable): Weight quantizater.
-            See more at `blueoil.nn.quantizations`.
+            See more at `blueoil.quantizations`.
         activation_quantizer_kwargs (dict): Kwargs for `activation_quantizer`.
         weight_quantizer (callable): Activation quantizater.
-            See more at `blueoil.nn.quantizations`.
+            See more at `blueoil.quantizations`.
         weight_quantizer_kwargs (dict): Kwargs for `weight_quantizer`.
 
     """
