+++ blueoil/quantizations/linear.py
