+++ blueoil/quantizations/ternary.py
