+++ blueoil/templates/lmnet/classification.tpl.py
@@ -28,7 +28,7 @@ from blueoil.pre_processor import (
     DivideBy255,
     PerImageStandardization
 )
-from blueoil.nn.quantizations import (
+from blueoil.quantizations import (
     binary_mean_scaling_quantizer,
     linear_mid_tread_half_quantizer,
 )
