+++ blueoil/templates/lmnet/object_detection.tpl.py
@@ -33,7 +33,7 @@ from blueoil.post_processor import (
     ExcludeLowScoreBox,
     NMS,
 )
-from blueoil.nn.quantizations import (
+from blueoil.quantizations import (
     binary_channel_wise_mean_scaling_quantizer,
     linear_mid_tread_half_quantizer,
 )
