+++ tests/unit/fixtures/configs/for_build_tfds_object_detection.py
@@ -32,7 +32,7 @@ from blueoil.post_processor import (
 from blueoil.data_augmentor import (
     FlipLeftRight,
 )
-from blueoil.nn.quantizations import (
+from blueoil.quantizations import (
     binary_mean_scaling_quantizer,
     linear_mid_tread_half_quantizer,
 )
