+++ tests/unit/networks_tests/keypoint_detection_tests/test_lm_single_pose_v1.py
@@ -28,7 +28,7 @@ from blueoil.pre_processor import (
     JointsToGaussianHeatmap
 )
 from blueoil.data_processor import Sequence
-from blueoil.nn.quantizations import (
+from blueoil.quantizations import (
     binary_channel_wise_mean_scaling_quantizer,
     linear_mid_tread_half_quantizer,
 )
