+++ tests/unit/networks_tests/object_detection_tests/test_yolo_v2_quantize.py
@@ -23,7 +23,7 @@ from blueoil.common import Tasks
 from blueoil.networks.object_detection.yolo_v2_quantize import YoloV2Quantize
 from blueoil.datasets.pascalvoc_2007 import Pascalvoc2007
 from blueoil.pre_processor import ResizeWithGtBoxes
-from blueoil.nn.quantizations import binary_channel_wise_mean_scaling_quantizer, linear_mid_tread_half_quantizer
+from blueoil.quantizations import binary_channel_wise_mean_scaling_quantizer, linear_mid_tread_half_quantizer
 from blueoil.utils.executor import prepare_dirs
 
 # Apply reset_default_graph() in conftest.py to all tests in this file.
