+++ tests/unit/test_quantizations.py
@@ -17,7 +17,7 @@ import numpy as np
 import pytest
 import tensorflow as tf
 
-from blueoil.nn.quantizations import (
+from blueoil.quantizations import (
     binary_channel_wise_mean_scaling_quantizer,
     binary_mean_scaling_quantizer,
     linear_mid_tread_half_quantizer,
